Dynamic Strands texture coordinates.
Polar coordinate and profile position all exposed to uniform particles.
Fixed UV seam artifacts.
Allow dynamic adjusting UV multiplier.
