+++ EvoEngine_Plugins/EcoSysLab/Internals/EcoSysLabResources/Shaders/Graphics/Mesh/DynamicStrandsRendering.mesh
@@ -69,25 +69,26 @@ void main(){
 	{
 		uint vertex_index = laneID + i * WORKGROUP_SIZE;
 		if (vertex_index >= TETRAHEDRON_VERTICES_SIZE) break;
-		vec3 vertex_position = uniform_particles[tet.indices[vertex_index]].position_t.xyz;
-		vec3 vertex_normal = normalize(uniform_particles[tet.indices[vertex_index]].normal_deg.xyz);
+		UniformParticle uniform_particle = uniform_particles[tet.indices[vertex_index]];
+		vec3 vertex_position = uniform_particle.position_t.xyz;
+		vec3 vertex_normal = normalize(uniform_particle.normal_deg.xyz);
 
 		ms_v_out[vertex_index].FragPos = vertex_position;
 		ms_v_out[vertex_index].Normal = vertex_normal;
-		ms_v_out[vertex_index].Tangent = uniform_particles[tet.indices[vertex_index]].tangent.xyz;
-		ms_v_out[vertex_index].TexCoord = vec2(uniform_particles[tet.indices[vertex_index]].tex_coord.x, uniform_particles[tet.indices[vertex_index]].tex_coord.y);
+		ms_v_out[vertex_index].Tangent = uniform_particle.tangent.xyz;
+		float tex_x = mod(uniform_particle.profile_polar_coordinate.y * u_multiplier / 3.1415926f, 2.f);
+		if(tex_x > 1.f){
+			tex_x = 2.f - tex_x;
+		}
+		float tex_y = float(uniform_particle.segment_index) * v_multiplier;
+		ms_v_out[vertex_index].TexCoord = vec2(tex_x, tex_y);
 		if(vertex_colors == 0){
 			ms_v_out[vertex_index].Color = vec4(0.5, 0.5, 0.5, 1.0);
 		} else if(vertex_colors == 1){
 			ms_v_out[vertex_index].Color = vec4(abs(vertex_normal.xyz), 1.0f);
 		} else if(vertex_colors == 2){
-			ms_v_out[vertex_index].Color = vec4(abs(uniform_particles[tet.indices[vertex_index]].tangent.xyz), 1.0f);
-		} else if(vertex_colors == 3){
-			ms_v_out[vertex_index].Color = vec4(mod(uniform_particles[tet.indices[vertex_index]].tex_coord.x, 1.0f),
-				mod(uniform_particles[tet.indices[vertex_index]].tex_coord.y, 1.0f),
-				mod(uniform_particles[tet.indices[vertex_index]].tex_coord.z, 1.0f), 1.0f);
+			ms_v_out[vertex_index].Color = vec4(abs(uniform_particle.tangent.xyz), 1.0f);
 		}
-
 		gl_MeshVerticesEXT[vertex_index].gl_Position = transform * vec4(vertex_position, 1.0);
 	}
