+++ EvoEngine_Plugins/EcoSysLab/Internals/EcoSysLabResources/Shaders/Graphics/Mesh/DynamicStrandsSpotLightShadowMap.mesh
@@ -23,14 +23,6 @@ layout(max_vertices = TETRAHEDRON_VERTICES_SIZE, max_primitives = TETRAHEDRON_TR
 const uint TETRAHEDRON_VERTEX_ITERATIONS    = ((TETRAHEDRON_VERTICES_SIZE	+ WORKGROUP_SIZE - 1) / WORKGROUP_SIZE);
 const uint TETRAHEDRON_PRIMITIVE_ITERATIONS = ((TETRAHEDRON_TRIANGLE_SIZE	+ WORKGROUP_SIZE - 1) / WORKGROUP_SIZE);
 
-layout(location = 0) out MS_V_OUT {
-	vec3 FragPos;
-	vec3 Normal;
-	vec3 Tangent;
-	vec2 TexCoord;
-	vec4 Color;
-} ms_v_out[];
-
 struct Task{
 	uint	baseID;
 	uint8_t	deltaIDs[EXT_TASK_WORK_GROUP_INVOCATIONS];
@@ -70,24 +62,7 @@ void main(){
 		uint vertex_index = laneID + i * WORKGROUP_SIZE;
 		if (vertex_index >= TETRAHEDRON_VERTICES_SIZE) break;
 		vec3 vertex_position = uniform_particles[tet.indices[vertex_index]].position_t.xyz;
-		vec3 vertex_normal = normalize(uniform_particles[tet.indices[vertex_index]].normal_deg.xyz);
-
-		ms_v_out[vertex_index].FragPos = vertex_position;
-		ms_v_out[vertex_index].Normal = vertex_normal;
-		ms_v_out[vertex_index].Tangent = uniform_particles[tet.indices[vertex_index]].tangent.xyz;
-		ms_v_out[vertex_index].TexCoord = vec2(uniform_particles[tet.indices[vertex_index]].tex_coord.x, uniform_particles[tet.indices[vertex_index]].tex_coord.y);
-		if(vertex_colors == 0){
-			ms_v_out[vertex_index].Color = vec4(0.5, 0.5, 0.5, 1.0);
-		} else if(vertex_colors == 1){
-			ms_v_out[vertex_index].Color = vec4(abs(vertex_normal.xyz), 1.0f);
-		} else if(vertex_colors == 2){
-			ms_v_out[vertex_index].Color = vec4(abs(uniform_particles[tet.indices[vertex_index]].tangent.xyz), 1.0f);
-		} else if(vertex_colors == 3){
-			ms_v_out[vertex_index].Color = vec4(mod(uniform_particles[tet.indices[vertex_index]].tex_coord.x, 1.0f),
-				mod(uniform_particles[tet.indices[vertex_index]].tex_coord.y, 1.0f),
-				mod(uniform_particles[tet.indices[vertex_index]].tex_coord.z, 1.0f), 1.0f);
-		}
-
+		
 		gl_MeshVerticesEXT[vertex_index].gl_Position = transform * vec4(vertex_position, 1.0);
 	}
